change predictor class into funcs
